Split typed actor forward declarations
