+++ libvast/test/system/eraser.cpp
@@ -28,6 +28,7 @@
 #include "vast/error.hpp"
 #include "vast/expression.hpp"
 #include "vast/ids.hpp"
+#include "vast/system/actors.hpp"
 #include "vast/system/archive.hpp"
 #include "vast/system/index.hpp"
 #include "vast/system/posix_filesystem.hpp"
