+++ libvast/vast/system/counter.hpp
@@ -17,6 +17,7 @@
 
 #include "vast/expression.hpp"
 #include "vast/ids.hpp"
+#include "vast/system/actors.hpp"
 #include "vast/system/query_processor.hpp"
 
 #include <unordered_map>
