+++ libvast/vast/system/disk_monitor.hpp
@@ -16,6 +16,7 @@
 #include "vast/fwd.hpp"
 
 #include "vast/path.hpp"
+#include "vast/system/actors.hpp"
 
 #include <caf/typed_event_based_actor.hpp>
