+++ libvast/vast/system/eraser.hpp
@@ -16,6 +16,7 @@
 #include "vast/fwd.hpp"
 
 #include "vast/ids.hpp"
+#include "vast/system/actors.hpp"
 #include "vast/system/query_processor.hpp"
 
 #include <string>
