+++ libvast/vast/system/index.hpp
@@ -21,6 +21,7 @@
 #include "vast/fbs/index.hpp"
 #include "vast/meta_index.hpp"
 #include "vast/system/accountant.hpp"
+#include "vast/system/actors.hpp"
 #include "vast/system/partition.hpp"
 #include "vast/uuid.hpp"
