+++ libvast/vast/system/posix_filesystem.hpp
@@ -16,6 +16,7 @@
 #include "vast/fwd.hpp"
 
 #include "vast/path.hpp"
+#include "vast/system/actors.hpp"
 #include "vast/system/filesystem_statistics.hpp"
 
 #include <caf/typed_event_based_actor.hpp>
