+++ libvast/vast/system/query_supervisor.hpp
@@ -16,6 +16,7 @@
 #include "vast/fwd.hpp"
 
 #include "vast/ids.hpp"
+#include "vast/system/actors.hpp"
 #include "vast/uuid.hpp"
 
 #include <caf/detail/unordered_flat_map.hpp>
